fix: resolve golangci-lint issues
- Fix errcheck in notification service (fire-and-forget delivery logging)
- Fix errcheck in rabbitmq subscriber (best effort ack)
- Fix hugeParam by passing ServiceDeps by pointer
- Fix gocyclo by extracting helper methods from SMTP sendWithTLS
- Fix gosec integer overflow by using uint64 for limit/offset params

+++ internal/repo/contracts.go
@@ -36,7 +36,7 @@ type (
 	NotificationRepo interface {
 		Store(ctx context.Context, n *notification.InAppNotification) error
 		GetByID(ctx context.Context, id uuid.UUID) (*notification.InAppNotification, error)
-		GetByUserID(ctx context.Context, userID uuid.UUID, limit, offset int) ([]notification.InAppNotification, error)
+		GetByUserID(ctx context.Context, userID uuid.UUID, limit, offset uint64) ([]notification.InAppNotification, error)
 		MarkAsRead(ctx context.Context, id uuid.UUID) error
 		MarkAllAsRead(ctx context.Context, userID uuid.UUID) error
 		GetUnreadCount(ctx context.Context, userID uuid.UUID) (int, error)
